Add filters to get the continents and countries data from states and display them in the components
